Add workflows for version bumping / releasing, CI
Wasn't originally planning on this, but added in additional CI workflows
to help with versioning from PRs, making the process less manual. This
will help with building the documentations as well.

Adds bump_version workflow and release-drafter for creating notes. Also
incorporates some CI changes that were already introduced in #192.