Update docker compose logging options
Docker by default saves a single log file in a privileged location:
/var/lib/docker/containers/<container_id>-json.log, which can be
accessed via the Docker commands or directly.

This commit updates the logging options to set the maximum file size to
be 300kB and a maximum of 10 files to be created. When this limit is
reached, the oldest file / content will be overwritten.