Add "develop" options to compose-dev.yaml
As of Docker compose v2.23.0, Docker compose has a "watch" command that
automates rebuilding or syncing of containers. This is cleaner for
deploying a production-like environment. If an older version of compose
is installed `docker compose up` still works - just requires a manual
restart of the service.

This PR just adds an additional develop option to watch for changes to
either the Python code or the environment variables. If either of these
change, the autobidsportal and rq service is rebuilt.