Add message to authorize bidsdump user
Currently, there is no mention anywhere of the need to authorize the
"bidsdump" user on the CFMM DICOM server. This small change adds
instructions for doing so (copied from old wiki) on index page, as well
as a reminder in the flash message after a valid new study form has been
submitted.