Fix gradcorrect trigger logic
There was a bug with the previous implementation, where if a study had
a custom ria url (which is any study valid study), gradcorrect would not
be triggered (for 7T studies). This commit / PR fixes this bug such that
it gradcorrect will be triggered if the study exists or if there is a
custom ria url.